Fix Expo Router conflict by renaming src/app to src/core
Expo automatically detects src/app as Expo Router directory and tries
to use file-based routing. Renamed to src/core to use traditional
React Navigation instead.
