+++ lepkehalo-rn/App.tsx
@@ -1,2 +1,2 @@
-// Re-export the main App from src/app
-export { default } from './src/app/App';
+// Re-export the main App from src/core
+export { default } from './src/core/App';
