+++ lepkehalo-rn/src/core/App.tsx
